added function autoSize
This will shrink or enlarge the image so that both the width or height is at least what was passed in.  So if you pass in an image that is 400x600 and you pass in the value of 800, it will enlarge the image to 800x1200.  If you have a picture that is 1600x1600 and pass in 800 you will get an image 800x800.